common: Increase precision of gauges
Although each sample has a precision of 1 ms, the average and standard
deviation can be provided at a higher resolution.

Target: trunk
Require-notes: yes
Require-book: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: https://rb.dcache.org/r/8266/

+++ modules/common/src/main/java/org/dcache/commons/stats/RequestExecutionTimeGauge.java
@@ -7,13 +7,13 @@ public interface RequestExecutionTimeGauge {
      * return average over the last period, and start new period
      * @return
      */
-    long resetAndGetAverageExecutionTime();
+    double resetAndGetAverageExecutionTime();
 
     /**
      * return average over the lifetime of the gauge
      * @return
      */
-    long getAverageExecutionTime();
+    double getAverageExecutionTime();
 
     /**
      * @return the RMS of executionTime
@@ -40,7 +40,7 @@ public interface RequestExecutionTimeGauge {
      */
     String getName();
 
-    long getStandardDeviation();
+    double getStandardDeviation();
 
     /**
      *
